Refactor window into app module and split phases
