+++ pomme-client/src/app/state_slot.rs
