+++ pomme-client/src/ui/menu/mod.rs
@@ -9,8 +9,7 @@ use std::time::Instant;
 
 use serde::{Deserialize, Serialize};
 
-use crate::window::DisplayMode;
-
+use crate::app::core::DisplayMode;
 use crate::renderer::pipelines::menu_overlay::{
     ICON_CHECK, ICON_CODE, ICON_COMMENT, ICON_GEAR, ICON_GLOBE, ICON_LINK, ICON_PAINTBRUSH,
     ICON_USER, MenuElement, SpriteId,
