--- pomme-client/src/window/game.rs
@@ -1,1 +1,0 @@
-pub struct GameState {}
